+++ packages/compiler-cli/ngcc/BUILD.bazel
@@ -25,6 +25,7 @@ ts_library(
         "//packages/compiler-cli/src/ngtsc/perf",
         "//packages/compiler-cli/src/ngtsc/reflection",
         "//packages/compiler-cli/src/ngtsc/scope",
+        "//packages/compiler-cli/src/ngtsc/sourcemaps",
         "//packages/compiler-cli/src/ngtsc/transform",
         "//packages/compiler-cli/src/ngtsc/translator",
         "//packages/compiler-cli/src/ngtsc/util",
@@ -36,8 +37,6 @@ ts_library(
         "@npm//dependency-graph",
         "@npm//magic-string",
         "@npm//semver",
-        "@npm//source-map",
-        "@npm//sourcemap-codec",
         "@npm//typescript",
     ],
 )
