+++ packages/compiler-cli/src/ngtsc/sourcemaps/BUILD.bazel
@@ -1,0 +1,18 @@
+load("//tools:defaults.bzl", "ts_library")
+
+package(default_visibility = ["//visibility:public"])
+
+ts_library(
+    name = "sourcemaps",
+    srcs = ["index.ts"] + glob([
+        "src/**/*.ts",
+    ]),
+    deps = [
+        "//packages/compiler-cli/src/ngtsc/file_system",
+        "//packages/compiler-cli/src/ngtsc/logging",
+        "@npm//@types/convert-source-map",
+        "@npm//@types/node",
+        "@npm//source-map",
+        "@npm//sourcemap-codec",
+    ],
+)
