+++ packages/compiler-cli/src/ngtsc/sourcemaps/README.md
@@ -1,0 +1,3 @@
+# Source-map handling
+
+Here there are classes for loading and merging source-maps.
