+++ packages/compiler-cli/src/ngtsc/sourcemaps/index.ts
@@ -1,0 +1,10 @@
+/**
+ * @license
+ * Copyright Google LLC All Rights Reserved.
+ *
+ * Use of this source code is governed by an MIT-style license that can be
+ * found in the LICENSE file at https://angular.io/license
+ */
+export {RawSourceMap} from './src/raw_source_map';
+export {Mapping, SourceFile} from './src/source_file';
+export {SourceFileLoader} from './src/source_file_loader';
