+++ packages/compiler-cli/src/ngtsc/sourcemaps/src/segment_marker.ts
