+++ packages/compiler-cli/src/ngtsc/sourcemaps/src/source_file_loader.ts
@@ -7,8 +7,8 @@
  */
 import {commentRegex, fromComment, mapFileCommentRegex} from 'convert-source-map';
 
-import {absoluteFrom, AbsoluteFsPath, FileSystem} from '../../../src/ngtsc/file_system';
-import {Logger} from '../../../src/ngtsc/logging';
+import {absoluteFrom, AbsoluteFsPath, FileSystem} from '../../file_system';
+import {Logger} from '../../logging';
 
 import {RawSourceMap} from './raw_source_map';
 import {SourceFile} from './source_file';
